docs(Editorial): moved to separate page
